Add modular package structure and pip installability
- Refactor monolithic instr_response.py into modular euvst_response package
- Add CLI interface with ASCII logo
- Add setup.py and pyproject.toml for pip installation
- Create console scripts: m-eclipses and meclipses
- Update README with installation and usage instructions
- Maintain backward compatibility with legacy imports
