+++ README.md
@@ -1,21 +1,52 @@
 # M-ECLIPSES: MSSL Emission Calculation and Line Intensity Prediction for SOLAR-C EUVST-SW
 
-```
-  __  __        ______ _____ _      _____ _____   _____ ______  _____ 
-  |  \/  |      |  ____/ ____| |    |_   _|  __ \ / ____|  ____|/ ____|
-  | \  / |______| |__ | |    | |      | | | |__) | (___ | |__  | (___  
-  | |\/| |______|  __|| |    | |      | | |  ___/ \___ \|  __|  \___ \ 
-  | |  | |      | |___| |____| |____ _| |_| |     ____) | |____ ____) |
-  |_|  |_|      |______\_____|______|_____|_|    |_____/|______|_____/ 
-```
-
 The M-ECLIPSES code (MSSL Emission Calculation and Line Intensity Prediction for SOLAR-C EUVST-SW) is used to forward model the performance of the ESA/MSSL short wavelength camera, part of the EUV spectrograph EUVST, on SOLAR-C.
 
 Contact: James McKevitt (jm2@mssl.ucl.ac.uk). License: Contact for permission to use.
 
 The instrument response generated by this code will be updated during instrument development, testing, and commissioning.
 
+## Installation
+
+### From PyPI (recommended)
+
+```bash
+pip install m-eclipses
+```
+
+### From source
+
+```bash
+git clone https://github.com/jamesmckevitt/solc_euvst_sw_response.git
+cd solc_euvst_sw_response
+pip install -e .
+```
+
 ## Usage Instructions
+
+### Command Line Interface
+
+After installation, you can run M-ECLIPSES from the command line:
+
+```bash
+# Show logo and basic usage
+m-eclipses
+
+# Run instrument response simulation
+m-eclipses --config ./run/input/config.yaml
+
+# Alternative short command
+meclipses --config ./run/input/config.yaml
+```
+
+### Python API
+
+You can also use M-ECLIPSES as a Python library:
+
+```python
+import euvst_response
+from euvst_response import AluminiumFilter, Detector_SWC, Telescope_EUVST
+```
 
 ### 1. Generate contribution functions for the desired emission lines
 
@@ -39,13 +70,13 @@ This step can require a lot of memory at full resolution. A fully synthesised at
 
 Run the instrument response function using:
 ```bash
-python instr_response.py --config ./run/input/config.yaml
+m-eclipses --config ./run/input/config.yaml
 ```
 
 This can be looped in bash using:
 ```bash
 for config in ./run/input/*.yaml; do
-  python instr_response.py --config "$config"
+  m-eclipses --config "$config"
 done
 ```
