+++ euvst_response/fitting.py
@@ -1,0 +1,139 @@
+"""
+Spectral fitting functions for Gaussian line profile analysis.
+"""
+
+from __future__ import annotations
+import warnings
+import numpy as np
+import astropy.units as u
+import astropy.constants as const
+from ndcube import NDCube
+from scipy.optimize import curve_fit, OptimizeWarning
+from joblib import Parallel, delayed
+from tqdm import tqdm
+from .utils import gaussian, tqdm_joblib
+
+
+def _guess_params(wv: np.ndarray, prof: np.ndarray) -> list:
+    """Guess initial parameters for Gaussian fit."""
+    back = prof.min()
+    prof_c = prof - back
+    prof_c[prof_c < 0] = 0
+    peak = prof_c.max()
+    centre = wv[np.nanargmax(prof_c)]
+    if peak == 0:
+        sigma = (wv.max() - wv.min()) / 10
+    else:
+        # Simple FWHM estimate
+        half_max = 0.5 * peak
+        indices = np.where(prof_c >= half_max)[0]
+        if len(indices) > 1:
+            fwhm = wv[indices[-1]] - wv[indices[0]]
+            sigma = fwhm / (2 * np.sqrt(2 * np.log(2)))
+        else:
+            sigma = (wv.max() - wv.min()) / 10
+    return [peak, centre, sigma, back]
+
+
+def _fit_one(wv: np.ndarray, prof: np.ndarray) -> np.ndarray:
+    """Fit single spectrum with Gaussian."""
+    p0 = _guess_params(wv, prof)
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", OptimizeWarning)
+        try:
+            popt, _ = curve_fit(gaussian, wv, prof, p0=p0)
+            return popt
+        except:
+            return np.array(p0)
+
+
+def fit_cube_gauss(signal_cube: NDCube, n_jobs: int = -1) -> u.Quantity:
+    """
+    Fit a Gaussian to every (slit x wavelength) spectrum.
+
+    Returns an array of shape (n_scan, n_slit, 4) with the parameters
+    [peak, centre, sigma, background] for each spatial pixel.
+    """
+
+    n_scan, n_slit, _ = signal_cube.shape
+    wv = signal_cube.axis_world_coords(2)[0].cgs  # wavelength axis (Quantity)
+
+    def _fit_block(spec_block: np.ndarray) -> np.ndarray:
+        results = np.empty((spec_block.shape[0], 4))
+        for i in range(spec_block.shape[0]):
+            results[i] = _fit_one(wv.value, spec_block[i])
+        return results
+
+    with tqdm_joblib(tqdm(total=n_scan, desc="Fit chunks", leave=False)):
+        results = Parallel(n_jobs=n_jobs)(
+            delayed(_fit_block)(signal_cube.data[i]) for i in range(n_scan)
+        )
+
+    # Stack to (n_scan, n_slit, 4) with units
+    arr = (np.stack(results, axis=0) * np.array([signal_cube.unit, wv.unit, wv.unit, signal_cube.unit], dtype=object))
+
+    return arr
+
+
+def velocity_from_fit(fit_arr: u.Quantity | np.ndarray, wl0: u.Quantity, n_jobs: int = -1) -> u.Quantity:
+    """
+    Convert fitted line centres to LOS velocity.
+    Works with either a Quantity array or an object-dtype array whose
+    elements are Quantities. Uses joblib.Parallel for speed.
+    """
+    centres_raw = fit_arr[..., 1]  # (n_scan, n_slit)
+    # Ensure we have a pure Quantity array
+    if isinstance(centres_raw, u.Quantity):
+        centres = centres_raw.to(wl0.unit)
+    else:  # object array of Quantity scalars
+        get_val = np.vectorize(lambda q: q.to_value(wl0.unit))
+        centres = u.Quantity(get_val(centres_raw), wl0.unit)
+
+    n_scan = centres.shape[0]
+
+    def _one_row(i):
+        return ((centres[i] - wl0) / wl0 * const.c).to(u.cm / u.s).value
+
+    with tqdm_joblib(tqdm(total=n_scan, desc="Velocity calc", leave=False)):
+        v_val = np.array(
+            Parallel(n_jobs=n_jobs)(
+                delayed(_one_row)(i) for i in range(n_scan)
+            )
+        )
+
+    v = v_val * (u.cm / u.s)
+    return v
+
+
+def width_from_fit(fit_arr: u.Quantity | np.ndarray, n_jobs: int = -1) -> u.Quantity:
+    """
+    Extract fitted line widths (sigma) from fit results.
+    """
+    widths_raw = fit_arr[..., 2]  # (n_scan, n_slit) - sigma is 3rd parameter
+    # Ensure we have a pure Quantity array
+    if isinstance(widths_raw, u.Quantity):
+        widths = widths_raw
+    else:  # object array of Quantity scalars
+        get_val = np.vectorize(lambda q: q.value)
+        get_unit = widths_raw.flat[0].unit  # Get unit from first element
+        widths = u.Quantity(get_val(widths_raw), get_unit)
+    
+    return widths
+
+
+def analyse(fits_all: u.Quantity | np.ndarray, v_true: u.Quantity, wl0: u.Quantity) -> dict:
+    """
+    Monte-Carlo velocity statistics given pre-computed ground truth.
+    """
+    v_all = velocity_from_fit(fits_all, wl0)
+    w_all = width_from_fit(fits_all)
+    return {
+        "v_mean": v_all.mean(axis=0),
+        "v_std":  v_all.std(axis=0),
+        "v_err":  v_true - v_all.mean(axis=0),
+        "v_samples": v_all,
+        "v_true":    v_true,
+        "w_mean": w_all.mean(axis=0),
+        "w_std":  w_all.std(axis=0),
+        "w_samples": w_all,
+    }
