+++ euvst_response/psf.py
@@ -1,0 +1,46 @@
+"""
+PSF handling functions for telescope and detector point spread functions.
+"""
+
+from __future__ import annotations
+from pathlib import Path
+import numpy as np
+import astropy.units as u
+from scipy.ndimage import zoom
+from scipy.signal import convolve2d
+
+
+def _load_psf_ascii(fname: Path, skip: int) -> np.ndarray:
+    """Load PSF data from ASCII file."""
+    return np.loadtxt(fname, skiprows=skip, encoding="utf-16 LE")
+
+
+def _resample_psf(psf: np.ndarray, res_in: u.Quantity, res_out: u.Quantity) -> np.ndarray:
+    """Resample PSF to different pixel scale."""
+    factor = (res_in / res_out).decompose().value
+    return zoom(psf, factor, order=1)
+
+
+def _combine_psfs(psf_focus: np.ndarray, psf_mesh: np.ndarray, crop: float = 0.99, size: int | None = None) -> np.ndarray:
+    """Convolve focus and mesh PSF and crop to given energy or size."""
+    psf = convolve2d(psf_focus, psf_mesh, mode="same")
+    if size is not None:
+        if size % 2 == 0:
+            size += 1
+        r0, c0 = np.array(psf.shape) // 2
+        half = size // 2
+        psf = psf[r0 - half : r0 + half + 1, c0 - half : c0 + half + 1]
+    else:
+        flat = psf.ravel()
+        idx = flat.argsort()[::-1]
+        csum = flat[idx].cumsum()
+        thr = flat[idx[np.searchsorted(csum, flat.sum() * crop)]]
+        flat[flat < thr] = 0
+        rows, cols = np.where(flat.reshape(psf.shape))
+        r0, r1 = rows.min(), rows.max()
+        c0, c1 = cols.min(), cols.max()
+        side = max(r1 - r0, c1 - c0) + 1
+        r0 = (r0 + r1) // 2 - side // 2
+        c0 = (c0 + c1) // 2 - side // 2
+        psf = psf[r0 : r0 + side, c0 : c0 + side]
+    return psf / psf.sum()
