+++ pyproject.toml
@@ -1,0 +1,59 @@
+[build-system]
+requires = ["setuptools>=45", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "m-eclipses"
+version = "1.0.0"
+authors = [
+    {name = "James McKevitt", email = "jm2@mssl.ucl.ac.uk"},
+]
+description = "MSSL Emission Calculation and Line Intensity Prediction for SOLAR-C EUVST-SW"
+readme = "README.md"
+license = {text = "Contact for permission"}
+requires-python = ">=3.8"
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Science/Research",
+    "Topic :: Scientific/Engineering :: Astronomy",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+]
+dependencies = [
+    "numpy",
+    "astropy",
+    "ndcube",
+    "specutils",
+    "scipy",
+    "matplotlib",
+    "joblib",
+    "tqdm",
+    "dill",
+    "pyyaml",
+    "reproject",
+]
+
+[project.optional-dependencies]
+dev = [
+    "pytest",
+    "black",
+    "flake8",
+    "mypy",
+]
+
+[project.scripts]
+m-eclipses = "euvst_response.cli:main"
+meclipses = "euvst_response.cli:main"
+
+[project.urls]
+Homepage = "https://github.com/jamesmckevitt/solc_euvst_sw_response"
+Repository = "https://github.com/jamesmckevitt/solc_euvst_sw_response"
+
+[tool.setuptools.packages.find]
+include = ["euvst_response*"]
+
+[tool.setuptools.package-data]
+euvst_response = ["data/**/*"]
